Backend/requirements.txt
gevent==23.9.1
urllib3==1.26.18
Flask==2.2.5
pefile==2023.2.7
pandas==2.2.1
joblib==1.3.2
typing_extensions==4.9.0
PyMuPDF==1.23.26
mysql-connector-python==8.0.15
python-dotenv==1.0.1
seaborn==0.13.2
scikit-learn==1.4.1.post1
oletools==0.60.1
pyexifinfo==0.4.0
python-magic==0.4.27